backend/requirements.txt
django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.0
Pillow>=10.0
requests>=2.31
gunicorn>=21.0
whitenoise>=6.5
